refactor(posts): move preview files to respective features
