+++ features/posts/common/presentation/src/androidMain/kotlin/com/parksupark/soomjae/features/posts/common/presentation/previews/CommentPreview.kt
@@ -1,0 +1,27 @@
+package com.parksupark.soomjae.features.posts.common.presentation.previews
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewParameter
+import com.parksupark.soomjae.core.presentation.designsystem.components.SoomjaeSurface
+import com.parksupark.soomjae.core.presentation.designsystem.theme.AppTheme
+import com.parksupark.soomjae.features.posts.common.presentation.components.CommentItem
+import com.parksupark.soomjae.features.posts.common.presentation.models.CommentUi
+import com.parksupark.soomjae.features.posts.common.presentation.previews.providers.CommentUiPreviewParameterProvider
+import kotlin.time.ExperimentalTime
+import kotlinx.collections.immutable.ImmutableList
+
+@OptIn(ExperimentalTime::class)
+@Preview
+@Composable
+private fun CommentPreview(
+    @PreviewParameter(CommentUiPreviewParameterProvider::class) comments: ImmutableList<CommentUi>,
+) {
+    AppTheme {
+        SoomjaeSurface {
+            CommentItem(
+                comment = comments[0],
+            )
+        }
+    }
+}
