+++ features/posts/community/presentation/src/androidMain/kotlin/com/parksupark/soomjae/features/posts/community/presentation/previews/providers/CommunityPostDetailUiPreviewParameterProvider.kt
@@ -1,6 +1,8 @@
-package com.parksupark.soomjae.features.posts.aggregate.presentation.previews.proviers
+package com.parksupark.soomjae.features.posts.community.presentation.previews.providers
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import com.parksupark.soomjae.features.posts.common.presentation.models.toUi
+import com.parksupark.soomjae.features.posts.common.presentation.previews.providers.CommentPreviewParameterData
 import com.parksupark.soomjae.features.posts.community.presentation.models.CommunityPostDetailUi
 import com.parksupark.soomjae.features.posts.community.presentation.models.toUi
 import kotlin.time.ExperimentalTime
@@ -8,7 +10,7 @@ import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toPersistentList
 
-internal class CommunityPostDetailUiPreviewParameterProvider : PreviewParameterProvider<ImmutableList<CommunityPostDetailUi>> {
+class CommunityPostDetailUiPreviewParameterProvider : PreviewParameterProvider<ImmutableList<CommunityPostDetailUi>> {
     override val values: Sequence<ImmutableList<CommunityPostDetailUi>> = sequenceOf(
         CommunityPostDetailUiPreviewParameterData.communityPostDetailUi,
     )
